refactor: move `isArtsyEmail` to `Utils`
